+++ src/Route/Generator/Generator.php
@@ -55,7 +55,7 @@ abstract class Generator implements GeneratorInterface
 	/**
 	 * @inheritdoc
 	 */
-	public function generate(array $routes)
+	public function generate(array $routes):array
 	{
 		foreach ($routes as $route) {
 			$this->mapRoute($route);