+++ src/Strategy/StdAppStrategy.php
@@ -26,7 +26,7 @@ class StdAppStrategy implements StrategyInterface
 	/**
 	 * @inheritdoc
 	 */
-	public function getHeader()
+	public function getHeader():array
 	{
 		return ["name"=>'Content-Type',"value"=>'text/html'];
 	}